refactor(server): config and rename middlewares
